Fix delete stations bug
