+++ examples/vehicle_sharing/run_freefloating.py
@@ -3,8 +3,7 @@ from mnms.generation.roads import generate_manhattan_road
 from mnms.mobility_service.vehicle_sharing import OnVehicleSharingMobilityService
 from mnms.tools.observer import CSVVehicleObserver, CSVUserObserver
 from mnms.generation.layers import generate_layer_from_roads, generate_grid_origin_destination_layer
-from mnms.graph.layers import SharedVehicleLayer
-from mnms.graph.multilayer_graph import MultiLayerGraph
+from mnms.graph.layers import SharedVehicleLayer, MultiLayerGraph
 from mnms.vehicles.veh_type import Bike
 from mnms.generation.demand import generate_random_demand
 from mnms.travel_decision.dummy import DummyDecisionModel
@@ -65,6 +64,6 @@ supervisor = Supervisor(mlgraph,
 supervisor.run(Time("07:00:00"),
                 Time("09:00:00"),
                 Dt(minutes=1),
-                10)
+                1)
