+++ src/mnms/generation/mlgraph.py
@@ -1,7 +1,7 @@
 from mnms.generation.roads import generate_manhattan_road
 from mnms.generation.layers import generate_layer_from_roads, generate_matching_origin_destination_layer
 from mnms.mobility_service.personal_vehicle import PersonalMobilityService
-from mnms.graph.layers import MultiLayerGraph
+from mnms.graph.abstract import MultiLayerGraph
 
 
 def generate_manhattan_passenger_car(n, link_length, resid="RES") -> MultiLayerGraph:
