+++ src/mnms/travel_decision/abstract.py
@@ -8,7 +8,7 @@ import numpy as np
 from numpy.linalg import norm as _norm
 
 from mnms.demand.user import User, Path
-from mnms.graph.multilayer_graph import MultiLayerGraph
+from mnms.graph.layers import MultiLayerGraph
 from mnms.mobility_service.personal_vehicle import PersonalMobilityService
 from mnms.log import create_logger
 from mnms.time import Time
